Initial idea for generating a test function for an interface
The idea here is that we could provide an
`Interfaces.testinterface(::Type{IT}, x)` function that implementors
could put in their test suite like
`Interfaces.testinterface(AbstractTable, MyTable())`, so they specify
the interface they're testing against, and an _instance_ of their
implementing type. In the `@interface` definition, a `@test begin ...
end` block can be provided that, at a minimum, takes an instance
variable name like `x` and checks whatever necessary for the interface.

I think we'll need to allow passing in other variables as well, like for
the `setindex!` case for `AbstractArray`.